backend/requirements.txt
google-api-python-client==2.99.0
scikit-learn~=1.0.2
numpy==1.24.4
pandas~=1.4.2
nltk~=3.7
pymystem3==0.2.0
pyyaml == 6.0.1
fastapi~=0.96.0
pydantic~=1.8
uvicorn~=0.22.0